server/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
mysql-connector-python==8.1.0
python-dotenv==1.0.0
gunicorn==20.1.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
requests==2.31.0